ETag management
When read fill in the ETag with the last write time utc.
When write check if the current last write time utc is the current ETag, if not it means someone wrote something in between. Throw ex to kill current activation.
When clear state check if the current last write time utc is the current ETag, if not it means someone wrote something in between. Throw ex to kill current activation.